% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate matrix-variate data}
\usage{
generate_data(
  n,
  K,
  p,
  gom,
  min_reps = 1,
  max_reps = 6,
  int = c(-5, 5),
  mix_prob = rep(1, K),
  complete = 0.05,
  rep_prob = rep(1, length(min_reps:max_reps)),
  method = "random"
)
}
\arguments{
\item{n}{Number of observations to generate}

\item{K}{Number of clusters}

\item{p}{Number of dimensions}

\item{gom}{Generalized overlap of Maitra}

\item{min_reps}{Minimum number of repetitions per observation}

\item{max_reps}{Maximum number of repetitions per observation}

\item{int}{Vector containing lower and upper bounds for the centers of each distribution}

\item{mix_prob}{Vector of mixing proportions for each cluster}

\item{complete}{Proportion of fully observed cases (with \code{max_reps} rows of
fully observed data) to be present for each cluster}

\item{rep_prob}{Vector of probabilities for each value of min_reps:max_reps
to occur in each observation}

\item{method}{Character string controlling how missing replicates are
introduced. \code{"random"} (default) drops replicates from the end of the
generated draws; \code{"lowest"} drops the replicates with the smallest values
for each dimension.}
}
\value{
A list, containing an array of data with observations on the 3rd axis
and a vector of class memberships
}
\description{
Generate matrix-variate data
}
\examples{
simulated_data <- generate_data(1000, 4, 10, 0.2)
res <- repclust(simulated_data$data, 4)
}
