\name{matclust-package}
\alias{matclust-package}
\alias{matclust}
\docType{package}
\title{
\packageTitle{matclust}
}
\description{
\packageDescription{matclust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matclust}
\packageIndices{matclust}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{matclust}

Maintainer: \packageMaintainer{matclust}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
