% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{run_em}
\alias{run_em}
\title{run_em}
\usage{
run_em(x, nclusters, iter_max = 101, tol = 0.001, init = "kmeans")
}
\value{
A list
}
\description{
A function to run the EM algorithm on matrix-variate data
}
\examples{
run_em()
}
