% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{repclust}
\alias{repclust}
\title{repclust}
\usage{
repclust(
  x,
  nclusters,
  iter_max = 101,
  tol = 0.001,
  init = "kmeans",
  params = NULL,
  emEM_args = list(nstarts = round(sqrt(nclusters * prod(dim(x)))), em_iter = 1, nbest =
    10)
)
}
\value{
A list
}
\description{
A function to run the EM algorithm on matrix-variate data
}
\examples{
repclust()
}
